pyproject.toml
[tool.poetry]
name = "pf-video-transcribe"
version = "0.1.0"
description = "Transcribe videos and create html pages with that"
authors = ["Gustavo Sverzut Barbieri <barbieri@profusion.mobi>"]
license = "MIT"
readme = "README.rst"
packages = [{include = "pf_video_transcribe"}]

[tool.poetry.scripts]
pf-video-transcribe = "pf_video_transcribe.__main__:cli"

[tool.poetry.dependencies]
python = "^3.11"
faster-whisper = "^0.6.0"
termcolor = "^2.3.0"
tqdm = "^4.65.0"
jinja2 = "^3.1.2"
ffmpeg-python = "^0.2.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.2"

[tool.mypy]
packages = ["pf_video_transcribe"]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.black]
line-length = 88
target-version = ['py311']

# disable isort, for IDEs that just default isort to be turned on, e.g. vscode.
# we use flake8-import-order for import sorting, using zimports to actually
# reformat code.  isort is nicer in many ways but doesn't have our
# "import *" fixer and also is not 100% compatible with flake8-import-order.
[tool.isort]
skip_glob=['*']

[tool.zimports]
black-line-length = 88
keep-unused-type-checking = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
